Add support for data-valid-email
In order to make it easier for the user, add a shortcut to test
the email regex pattern so the user doesn't have to pass in a
pattern when trying to validate.  Refactor the library to add a
method that handles the regex pattern testing and returns a
boolean. Add test for data-valid-email.

Fix #4